128bit NTT

+++ hexl/ntt/ntt-internal.hpp
@@ -32,10 +32,10 @@ namespace hexl {
 /// @param[in] output_mod_factor Upper bound for result; result must be in [0,
 /// output_mod_factor * q)
 void ForwardTransformToBitReverseRadix2(
-    uint64_t* result, const uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* root_of_unity_powers,
-    const uint64_t* precon_root_of_unity_powers, uint64_t input_mod_factor = 1,
-    uint64_t output_mod_factor = 1);
+    uint128_t* result, const uint128_t* operand, size_t n, uint128_t modulus,
+    const uint128_t* root_of_unity_powers,
+    const uint128_t* precon_root_of_unity_powers, uint128_t input_mod_factor = 1,
+    uint128_t output_mod_factor = 1);
 
 /// @brief Radix-4 native C++ NTT implementation of the forward NTT
 /// @param[out] result Output data. Overwritten with NTT output
@@ -52,10 +52,10 @@ void ForwardTransformToBitReverseRadix2(
 /// @param[in] output_mod_factor Upper bound for result; result must be in [0,
 /// output_mod_factor * q)
 void ForwardTransformToBitReverseRadix4(
-    uint64_t* result, const uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* root_of_unity_powers,
-    const uint64_t* precon_root_of_unity_powers, uint64_t input_mod_factor = 1,
-    uint64_t output_mod_factor = 1);
+    uint128_t* result, const uint128_t* operand, size_t n, uint128_t modulus,
+    const uint128_t* root_of_unity_powers,
+    const uint128_t* precon_root_of_unity_powers, uint128_t input_mod_factor = 1,
+    uint128_t output_mod_factor = 1);
 
 /// @brief Reference forward NTT which is written for clarity rather than
 /// performance
@@ -66,8 +66,8 @@ void ForwardTransformToBitReverseRadix4(
 /// @param[in] root_of_unity_powers Powers of 2n'th root of unity in F_q. In
 /// bit-reversed order
 void ReferenceForwardTransformToBitReverse(
-    uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* root_of_unity_powers);
+    uint128_t* operand, uint128_t n, uint128_t modulus,
+    const uint128_t* root_of_unity_powers);
 
 /// @brief Reference inverse NTT which is written for clarity rather than
 /// performance
@@ -78,8 +78,8 @@ void ReferenceForwardTransformToBitReverse(
 /// @param[in] inv_root_of_unity_powers Powers of inverse 2n'th root of unity in
 /// F_q. In bit-reversed order.
 void ReferenceInverseTransformFromBitReverse(
-    uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* inv_root_of_unity_powers);
+    uint128_t* operand, uint128_t n, uint128_t modulus,
+    const uint128_t* inv_root_of_unity_powers);
 
 /// @brief Radix-2 native C++ NTT implementation of the inverse NTT
 /// @param[out] result Output data. Overwritten with NTT output
@@ -96,10 +96,10 @@ void ReferenceInverseTransformFromBitReverse(
 /// @param[in] output_mod_factor Upper bound for result; result must be in [0,
 /// output_mod_factor * q)
 void InverseTransformFromBitReverseRadix2(
-    uint64_t* result, const uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* inv_root_of_unity_powers,
-    const uint64_t* precon_inv_root_of_unity_powers,
-    uint64_t input_mod_factor = 1, uint64_t output_mod_factor = 1);
+    uint128_t* result, const uint128_t* operand, size_t n, uint128_t modulus,
+    const uint128_t* inv_root_of_unity_powers,
+    const uint128_t* precon_inv_root_of_unity_powers,
+    uint128_t input_mod_factor = 1, uint128_t output_mod_factor = 1);
 
 /// @brief Radix-4 native C++ NTT implementation of the inverse NTT
 /// @param[out] result Output data. Overwritten with NTT output
@@ -116,10 +116,10 @@ void InverseTransformFromBitReverseRadix2(
 /// @param[in] output_mod_factor Upper bound for result; result must be in [0,
 /// output_mod_factor * q)
 void InverseTransformFromBitReverseRadix4(
-    uint64_t* result, const uint64_t* operand, uint64_t n, uint64_t modulus,
-    const uint64_t* root_of_unity_powers,
-    const uint64_t* precon_root_of_unity_powers, uint64_t input_mod_factor = 1,
-    uint64_t output_mod_factor = 1);
+    uint128_t* result, const uint128_t* operand, size_t n, uint128_t modulus,
+    const uint128_t* root_of_unity_powers,
+    const uint128_t* precon_root_of_unity_powers, uint128_t input_mod_factor = 1,
+    uint128_t output_mod_factor = 1);
 
 }  // namespace hexl
 }  // namespace intel
